fix: layout shift caused by Dropdownmenu and make it to a separate component
